refactor: share proposal building between z_sendmany and pczt_create
Extract the propose/validate pipeline that z_sendmany and the PCZT methods
have in common into a shared payments::propose_and_check helper: the
privacy-policy validation loop, propose_transfer with the standard change
strategy and input selector, privacy-policy enforcement, and the
Orchard-action-limit check.

Move the pieces that go with it out of z_send_many.rs and into payments.rs,
next to that helper: spend_policy_for, which confines a transparent source
to its own UTXOs, and the minconf-to-ConfirmationsPolicy mapping, now
confirmations_policy_for_minconf. Route z_sendmany's inline privacy-policy
parsing through the existing parse_privacy_policy helper rather than
carrying a byte-for-byte copy of it.

Also adopt the shared FTL keys at these pre-existing message sites, so each
message has one definition rather than copies that can drift. Only messages
the new methods will share are converted, and only where the conversion is
unambiguous. z_sendmany retains several of its own: the fee-field check; its
from-address error, which is not in fact a duplicate since it also accepts
the literal ANY_TADDR; and the two chained “Failed to propose transaction”
arms, which want analysis before being localized — the first reports a
task-join failure rather than a proposal one, and the call they wrap creates
transactions rather than proposing them.

Finally, collapse the per-branch privacy checks onto a single classifier.
enforce_privacy_policy tested each branch with a differently-named allow_*
predicate, every one of which was is_compatible_with of the corresponding
policy. step_privacy_requirement now returns that policy directly, paired with
the error to report when the caller's does not reach it, so the check is driven
by data rather than by a name per branch. The four allow_* wrappers with no
remaining callers go with it.
